Fix Remove TerminalApiRegionTest, add unit test in RegionTest
